% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fu.r
\name{cond_terms}
\alias{cond_terms}
\title{Get the Conditional Terms of a Formula}
\usage{
cond_terms(form, data = NULL)
}
\arguments{
\item{data}{an associated data set. Default NULL (no dataset)}

\item{formula}{the formula to extract the right hand terms from.}
}
\description{
Get the Conditional Terms of a Formula
}
