% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fu.r
\name{expand_formula}
\alias{expand_formula}
\title{Expand a Formula with a dot}
\usage{
expand_formula(x, form)
}
\arguments{
\item{x}{the data set.}

\item{form}{the formula.}
}
\description{
Expand a Formula with a dot
}
