% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fu.r
\name{extract_vars}
\alias{extract_vars}
\title{Extract the Variables from a Character Expression}
\usage{
extract_vars(str_expr)
}
\arguments{
\item{str_expr}{the string expression}
}
\description{
Extract the Variables from a Character Expression
}
