% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fu.r
\name{form_desc}
\alias{form_desc}
\title{Get a Description of a formula Object}
\usage{
form_desc(x, form)
}
\arguments{
\item{x}{a data.frame.}

\item{form}{a formula object.}
}
\description{
Get a Description of a formula Object
}
