% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fu.r
\name{indep_terms}
\alias{indep_terms}
\title{Get the Independent Terms of a Formula}
\usage{
indep_terms(form, data = NULL)
}
\arguments{
\item{data}{an associated data set. Default NULL (no dataset)}

\item{formula}{the formula to extract the right hand terms from.}
}
\description{
Get the Independent Terms of a Formula
}
