% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fu.r
\name{lh_terms}
\alias{lh_terms}
\title{Get the Left Hand Terms}
\usage{
lh_terms(form, data = NULL)
}
\arguments{
\item{data}{an associated data set. Default NULL (no dataset)}

\item{formula}{the formula to extract the right hand terms from.}
}
\description{
Get the Left Hand Terms
}
