% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fu.r
\name{make_formula}
\alias{make_formula}
\title{Make a formula}
\usage{
make_formula(dep_vars, indep_vars, cond_vars = NULL)
}
\arguments{
\item{dep_vars}{the dependent variables.}

\item{indep_vars}{the independent variables.}

\item{cond_vars}{the conditional variables. (Default NULL - none)}
}
\description{
Make a formula
}
