% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fu.r
\name{make_variable_desc}
\alias{make_variable_desc}
\title{Make a description of the variables in the formula.}
\usage{
make_variable_desc(x, form)
}
\arguments{
\item{x}{a data.frame containing the variable samples.}

\item{form}{the formula describing the model.}
}
\description{
Make a description of the variables in the formula.
}
