% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fu.r
\name{rh_terms}
\alias{rh_terms}
\title{Get the Right Hand Terms of a Formula}
\usage{
rh_terms(form, data = NULL)
}
\arguments{
\item{data}{an associated data set. Default NULL (no dataset)}

\item{formula}{the formula to extract the right hand terms from.}
}
\description{
Get the Right Hand Terms of a Formula
}
